% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dat.fleiss1971}
\alias{dat.fleiss1971}
\title{Fleiss (1971) psychiatric diagnoses dataset}
\format{
A 30 x 5 integer matrix.
}
\source{
Fleiss, J. L. (1971). Measuring nominal scale agreement among many
  raters. \emph{Psychological Bulletin}, 76(5), 378-382.
}
\usage{
dat.fleiss1971
}
\description{
Counts-format rating data from Fleiss (1971), n = 30 subjects, six raters,
five categories. Each row is one subject; each column counts the number of
raters who assigned that subject to that category.
}
\examples{
kappa_counts(dat.fleiss1971, estimator = "pairwise")
}
\keyword{datasets}
