% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unified_kappa_continuous_rcpp}
\alias{unified_kappa_continuous_rcpp}
\title{Unified C++ Backend for Continuous Data Kappa}
\usage{
unified_kappa_continuous_rcpp(x_r, method, weight_type)
}
\description{
Unified C++ Backend for Continuous Data Kappa
}
\keyword{internal}
