% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unified_kappa_counts_rcpp}
\alias{unified_kappa_counts_rcpp}
\title{Unified C++ Backend for Counts Data Kappa}
\usage{
unified_kappa_counts_rcpp(x_r, r, method, weight_type, values, options)
}
\description{
Unified C++ Backend for Counts Data Kappa
}
\keyword{internal}
