% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{misskappa-package}
\alias{misskappa}
\alias{misskappa-package}
\title{misskappa: Agreement Coefficients with Missing Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimation and inference for weighted agreement coefficients (Cohen, Fleiss, Conger, Brennan-Prediger) with arbitrary numbers of raters and arbitrary pairwise loss functions, plus coefficient alpha for scored categorical and continuous item batteries. Supports incomplete data under MCAR (available-case, IPW) and MAR (saturated categorical or normal FIML / EM), with sandwich standard errors. Wraps the standalone C++17 'misskappa' library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jonasmoss.github.io/misskappa}
  \item Report bugs at \url{https://github.com/JonasMoss/misskappa/issues}
}

}
\author{
\strong{Maintainer}: Jonas Moss \email{jonas.moss.statistics@gmail.com} (\href{https://orcid.org/0000-0002-6876-6964}{ORCID})

Authors:
\itemize{
  \item Jonas Moss \email{jonas.moss.statistics@gmail.com} (\href{https://orcid.org/0000-0002-6876-6964}{ORCID})
}

}
\keyword{internal}
